refactor(asset): modularise asset/image
